Moved code to YARP 2.3.61 / ACE 6.2.3; note that classes have been moved or removed in YARP and that the 'close' problem is less frequent.

+++ Parser/MoMeBaseMatcher.cpp
@@ -76,11 +76,11 @@ using namespace MoAndMe::Parser;
 # pragma mark Class methods
 #endif // defined(__APPLE__)
 
-int BaseMatcher::SkipWhitespace(const yarp::os::ConstString & inString,
-                                const int                     inLength,
-                                const int                     startPos)
+size_t BaseMatcher::SkipWhitespace(const yarp::os::ConstString & inString,
+                                   const size_t                  inLength,
+                                   const size_t                  startPos)
 {
-    int result = startPos;
+    size_t result = startPos;
     
     try
     {
