+++ Parser/MoMeMatchFieldWithValues.h
@@ -91,9 +91,9 @@ namespace MoAndMe
              @param validator A function that returns @c true if the field name is valid and @c false otherwise.
              @returns A non-null matcher if the string would be a valid value and @c NULL otherwise. */
             static MatchFieldWithValues * CreateMatcher(const yarp::os::ConstString & inString,
-                                                        const int                     inLength,
-                                                        const int                     startPos,
-                                                        int &                         endPos,
+                                                        const size_t                  inLength,
+                                                        const size_t                  startPos,
+                                                        size_t &                      endPos,
                                                         BaseNameValidator *           validator = NULL);
             
         protected:
